Removed links to empty TOC nodes on the index page of the new theme (v3)
