+++ source/_themes/wazuh_doc_theme_v3/src/js-source/components/in-release-redirections.js
@@ -6,9 +6,10 @@ if (rootLocation[rootLocation.length - 1] == '/') {
 }
 rootLocation = rootLocation.split(currentPath)[0];
 if ( redirectSameRelease && Object.keys(redirectSameRelease).length > 0 ) {
-  const newPath = redirectSameRelease[DOCUMENTATION_OPTIONS.VERSION][currentPath];
-  if ( newPath ) {
-    window.location.replace(rootLocation + newPath);
+  if (redirectSameRelease[DOCUMENTATION_OPTIONS.VERSION]) {
+    const newPath = redirectSameRelease[DOCUMENTATION_OPTIONS.VERSION][currentPath];
+    if ( newPath ) {
+      window.location.replace(rootLocation + newPath);
+    }
   }
 }
-
