+++ source/conf.py
@@ -14,6 +14,7 @@ import re
 import shlex
 import datetime
 import time
+import json
 from requests.utils import requote_uri
 
 # If extensions (or modules to document with autodoc) are in another directory,
@@ -521,6 +522,37 @@ custom_replacements = {
     "|ELASTICSEARCH_ELK_LATEST_DOCKER|" : "7.10.2",
 }
 
+# -- Customizations ---------------------------------------------------------
+# 
+# ## emptyTocNodes ##
+emptyTocNodes = json.dumps([
+  'style-guide/index',
+  'style-guide/text',
+  'style-guide/components',
+  'amazon/configuration/index',
+  'compliance',
+  'containers',
+  'deployment',
+  'development/index',
+  'docker-monitor/index',
+  'installation-guide/elasticsearch-cluster/index',
+  'installation-guide/wazuh-cluster/index',
+  'installation-guide/upgrading/legacy/index',
+  'installation-guide/packages-list/linux/linux-index',
+  'installation-guide/packages-list/solaris/solaris-index',
+  'monitoring',
+  'migration-guide/index',
+  'user-manual/index',
+  'user-manual/agents/index',
+  'user-manual/agents/remove-agents/index',
+  'user-manual/agents/listing/index',
+  'user-manual/kibana-app/reference/index',
+  'user-manual/ruleset/ruleset-xml-syntax/index',
+  'installation-guide/distributed-deployment/step-by-step-installation/elasticsearch-cluster/index',
+  'installation-guide/distributed-deployment/step-by-step-installation/wazuh-cluster/index',
+  'user-manual/capabilities/active-response/ar-use-cases/index',
+])
+
 # -- Setup -------------------------------------------------------------------
 
 compilation_time = str(time.time())
@@ -738,6 +770,7 @@ html_context = {
     "production": production,
     "apiURL": apiURL,
     "compilation_ts": compilation_time,
+    "empty_toc_nodes": emptyTocNodes,
     "is_latest_release": is_latest_release
 }
 sphinx_tabs_nowarn = True
